Project to Analyse and Predict Breast Cancer
Breast cancer is the most common invasive cancer in women and the second leading cause of cancer death in women after lung cancer. Advances in screening and treatment for breast cancer have improved survival rates dramatically since 1989. Early screening, leading to detection has, from statistics, led to the increase in survival rates. There is a possibility of detecting cancer at an early stage, based on features collected from the patients. Breast cancer can occur in women and rarely in men.
So the objective of this project is  the possibility of detecting cancer at an early stage, based on features collected from the patient. To also analyse the data and build a model that would detect cancerous cells with good accuracy